First stab at generating MASC-PCR primers for specific variants.
This new code attempts to re-use the existing code as much as possible.
Thus the pipeline is setup to create a modified genome and then reuse
Mauve alignment and other functions where possible.

It should be possible to refactor this to get rid of the need for Mauve
by manually building the lookup tables (LUTs).